Refactor drawer
