+++ src/AppRoute/index.js
@@ -1,8 +1,5 @@
 import React, { Component } from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import { connect } from 'react-redux';
-import PropTypes from 'prop-types';
-import { closeDrawer } from '../actions/ui';
 import TodoListScreen from '../screens/TodoListScreen';
 import CreateTodoScreen from '../screens/CreateTodoScreen';
 import ArchiveListScreen from '../screens/ArchiveListScreen';
@@ -20,31 +17,11 @@ class AppRoute extends Component {
             <Route exact path='/create' component={CreateTodoScreen} />
           </Switch>
           <SnackBar />
-          <Drawer
-            isOpen={this.props.isDrawerOpen}
-            onClosePress={this.props.closeDrawer}
-          />
+          <Drawer />
         </div>
       </BrowserRouter>
     );
   }
 }
 
-const mapStateToProps = ({ drawer }) => {
-  return {
-    isDrawerOpen: drawer.isOpen
-  };
-};
-
-const mapDispatchToProps = (dispatch) => {
-  return {
-    closeDrawer: () => dispatch(closeDrawer())
-  };
-};
-
-AppRoute.propTypes = {
-  isDrawerOpen: PropTypes.bool.isRequired,
-  closeDrawer: PropTypes.func.isRequired
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(AppRoute);
+export default AppRoute;
